feat: monitor config per user and allow setting wallpaper too
refactors monitor module and moves it into home manager modules.
most monitor configuration is used by home manager modules so it makes sense to put it there.
this allows users to configure monitors per host.
also added a wallpaper option to monitor module.
since wallpaper programs make use of monitor name/plug to figure out where to put the wallpaper
it makes sense to put it there so it doesn't have to be repeated in a seperate option

Though i would prefer to only use monitor name/desc for configuration. hyprpaper and hyprlock don't seem to work with those for me and require plug to be set.